+++ week04/src/pages/MovieDetail.tsx
@@ -1,71 +1,26 @@
-import axios from "axios";
-import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 
 import { CreditsResponse, MovieDetailResponse } from "../types/movie";
 import MovieDetailCredit from "../components/MovieDetailCredit";
 import Loading from "../components/Loading";
+import useApiRequest from "../../../week04/src/hooks/useApiRequest";
 
 const MovieDetail = () => {
-  const [data, setData] = useState<MovieDetailResponse | null>(null);
-  const [creditData, setCreditData] = useState<CreditsResponse | null>(null);
-
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState(false);
-
   const { id } = useParams<{ id: string }>();
+  const { data, loading, error } = useApiRequest<MovieDetailResponse>(
+    `https://api.themoviedb.org/3/movie/${id}`
+  );
+  const {
+    data: creditData,
+    loading: creditLoading,
+    error: creditError,
+  } = useApiRequest<CreditsResponse>(`https://api.themoviedb.org/3/movie/${id}/credits`);
 
-  useEffect(() => {
-    const fetchMovieDetail = async () => {
-      if (!id) {
-        return;
-      }
-
-      try {
-        const { data } = await axios.get<MovieDetailResponse>(
-          `https://api.themoviedb.org/3/movie/${id}`,
-          {
-            headers: {
-              Authorization: `Bearer ${import.meta.env.VITE_API_ACCESS_TOKEN}`,
-            },
-          }
-        );
-        setData(data);
-      } catch {
-        setError(true);
-      }
-    };
-
-    const fetchMovieCredits = async () => {
-      if (!id) {
-        return;
-      }
-      try {
-        const { data } = await axios.get<CreditsResponse>(
-          `https://api.themoviedb.org/3/movie/${id}/credits`,
-          {
-            headers: {
-              Authorization: `Bearer ${import.meta.env.VITE_API_ACCESS_TOKEN}`,
-            },
-          }
-        );
-        setCreditData(data);
-      } catch {
-        setError(true);
-      }
-    };
-
-    setLoading(true);
-    fetchMovieDetail();
-    fetchMovieCredits();
-    setLoading(false);
-  }, [id]);
-
-  if (error) {
+  if (error || creditError) {
     return <div className="text-red-500">Error occurred</div>;
   }
 
-  if (loading) {
+  if (loading || creditLoading) {
     return <Loading />;
   }
